Cleanup Storybook
