--- packages/common/.storybook/addons.js
@@ -1,2 +1,0 @@
-require('@storybook/addon-actions/register');
-require('@storybook/addon-knobs/register');
