+++ packages/common/.storybook/addons.ts
@@ -1,0 +1,2 @@
+import '@storybook/addon-actions/register';
+import '@storybook/addon-knobs/register';
