+++ packages/common/package.json
@@ -57,6 +57,7 @@
     "@types/lodash": "^4.14.123",
     "@types/react": "^16.8.12",
     "@types/react-icons": "^2.2.7",
+    "@types/storybook__addon-actions": "^3.4.3",
     "@types/storybook__addon-knobs": "^5.0.3",
     "@types/storybook__react": "^4.0.2",
     "@types/styled-components": "^4.1.13",
