+++ packages/common/src/stories/decorators/ThemeDecorator.tsx
@@ -1,6 +1,7 @@
-import * as React from 'react';
+import React from 'react';
 import { ThemeProvider } from 'styled-components';
-import theme from '../theme';
+
+import theme from '../../theme';
 
 export const ThemeDecorator = (fn: () => JSX.Element) => (
   <ThemeProvider theme={theme}>{fn()}</ThemeProvider>
