+++ packages/common/src/stories/decorators/index.ts
@@ -1,0 +1,2 @@
+export * from './ThemeDecorator';
+export * from './withGlobal';
